ragging-detection-system/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
tensorflow==2.19.0
numpy==2.1.3
pillow==11.1.0
python-multipart==0.0.20
opencv-python==4.11.0.86
requests==2.32.3
